Refactored code handle_admin command is now in adminbot.py instead of functions.py
