+++ functions.py
@@ -1057,348 +1057,6 @@ class OrcaBot(AdminBot, FunBot, TeleBot):
         """Main bot function to run"""
         TeleBot.main(self)
 
-    def handle_admin(self, update, context, keywords, command=''):
-        """Handle the admin commands after /admin"""
-        try:
-            if not command:
-                command=keywords.pop(0)
-
-            if command in ['setpin','setstatus','forcereturn']: #bike commands
-                bike_name  = keywords.pop(0) # set the bike_name as name
-                bikes_data = self.get_bikes()
-                bike = bikes_data.get(bike_name, None)
-                if bike is None:
-                    context.bot.send_message(
-                        chat_id=update.effective_chat.id,
-                        text=f'No such bike, {bike}, exists. Please try again with current /bikes'
-                    )
-                    return
-
-            elif command in ['topup','deduct','setcredit','user','ban','unban']: # user commands
-                username = keywords.pop(0) #set the username as name
-                all_users = self.get_user_table()
-                chat_id = all_users.get(username, None)
-                if chat_id is not None:
-                    with open(f'users/{chat_id}.json', 'r') as f:
-                        user_data = json.load(f)
-                else:
-                    context.bot.send_message(
-                        chat_id=update.effective_chat.id,
-                        text=f'Specified user is not found! Please ask @{username} to create an account first.')
-                    return
-            else: # stats and logs commands
-                pass
-
-            if command == "topup":
-                """Topup specific amount to user credits"""
-                number = keywords.pop(0)
-                user_data['credits'] = user_data.get('credits', 0)
-                initial_amt = user_data.get('credits', 0)
-                user_data['credits'] += int(number)
-                final_amt = user_data.get('credits')
-
-                #update user data
-                user_data['finance'] = user_data.get('finance',[])
-                f_log = {
-                    'type':'admin',
-                    'time':self.now().strftime("%Y/%m/%d, %H:%M:%S"),
-                    'initial':initial_amt,
-                    'change': int(number),
-                    'final': final_amt
-                }
-                user_data['finance'].append(f_log)
-                self.update_user(user_data)
-
-                finance_log=[
-                    username,
-                    self.now().strftime("%Y/%m/%d, %H:%M:%S"),
-                    initial_amt, int(number), final_amt,
-                    update.message.from_user.username
-                ]
-                self.update_finance_log(finance_log)
-                context.bot.send_message(
-                    chat_id=update.effective_chat.id,
-                    text=f'Top-up successful! @{username} now has {user_data["credits"]} credits.'
-                )
-
-            elif command == "deduct":
-                """Deduct specific amount from user credits"""
-                number = keywords.pop(0)
-                user_data['credits'] = user_data.get('credits', 0)
-                initial_amt = user_data.get('credits', 0)
-                user_data['credits'] -= int(number)
-                final_amt = user_data.get('credits')
-
-                #update user data
-                user_data['finance'] = user_data.get('finance',[])
-                f_log = {
-                    'type':'admin',
-                    'time':self.now().strftime("%Y/%m/%d, %H:%M:%S"),
-                    'initial':initial_amt,
-                    'change': -int(number),
-                    'final': final_amt
-                }
-                user_data['finance'].append(f_log)
-                self.update_user(user_data)
-
-                finance_log=[
-                    username,
-                    self.now().strftime("%Y/%m/%d, %H:%M:%S"),
-                    initial_amt, -int(number), final_amt,
-                    update.message.from_user.username
-                ]
-                self.update_finance_log(finance_log)
-                context.bot.send_message(
-                    chat_id=update.effective_chat.id,
-                    text=f'Deducted successfully! @{username} now has {user_data["credits"]} credits.'
-                )
-
-            elif command == "setcredit":
-                """Set user credits to specified amount."""
-                number = keywords.pop(0)
-                user_data['credits'] = user_data.get('credits', 0)
-                initial_amt = user_data.get('credits', 0)
-                change_amt = int(number) - initial_amt
-                user_data['credits'] = int(number)
-
-                #update user data
-                user_data['finance'] = user_data.get('finance',[])
-                f_log = {
-                    'type':'admin',
-                    'time':self.now().strftime("%Y/%m/%d, %H:%M:%S"),
-                    'initial':initial_amt,
-                    'change': change_amt,
-                    'final': int(number)
-                }
-                user_data['finance'].append(f_log)
-                self.update_user(user_data)
-
-                finance_log=[
-                    username,
-                    self.now().strftime("%Y/%m/%d, %H:%M:%S"),
-                    initial_amt, change_amt, int(number),
-                    update.message.from_user.username
-                ]
-                self.update_finance_log(finance_log)
-                context.bot.send_message(
-                    chat_id=update.effective_chat.id,
-                    text=f'Setting was successful! @{username} now has {user_data["credits"]} credits.'
-                )
-
-            elif command == "user":
-                text=f'@{user_data["username"]} has {user_data.get("credits",0)} credits left.\n'
-                text+=f'User has been renting {user_data["bike_name"]} since {user_data["status"]}' if user_data.get("bike_name") else "User is not renting currently."
-                context.bot.send_message(
-                    chat_id=update.effective_chat.id,
-                    text=text)
-
-            elif command == "ban":
-                """Ban a user"""
-                is_ban = user_data.get('is_ban')
-                if not is_ban:
-                    user_data['is_ban'] = True
-                    self.update_user(user_data)
-                    context.bot.send_message(
-                        chat_id=update.effective_chat.id,
-                        text=f'@{username} is now BANNED.'
-                    )
-                else:
-                    context.bot.send_message(
-                        chat_id=update.effective_chat.id,
-                        text=f'@{username} was already banned.'
-                    )
-
-            elif command == "unban":
-                """Unban a user"""
-                is_ban = user_data.get('is_ban')
-                if is_ban:
-                    user_data['is_ban'] = None
-                    self.update_user(user_data)
-                    context.bot.send_message(
-                        chat_id=update.effective_chat.id,
-                        text=f'@{username} is now UNBANNED.'
-                    )
-                else:
-                    context.bot.send_message(
-                        chat_id=update.effective_chat.id,
-                        text=f'@{username} was not banned'
-                    )
-
-            elif command == "forcereturn":
-                username = bike.get('username',"")
-                if username:
-                    user_data = self.get_user(username=username)
-                    status = user_data.get('status')
-                    if not status:
-                        return context.bot.send_message(
-                            chat_id=update.effective_chat.id,
-                            text=f"Something seems wrong... The bike {bike_name} tagged to user @{username}, but user is not renting???"
-                        )
-                    diff = self.now() - datetime.datetime.fromisoformat(status)
-                    if diff.days:
-                        strdiff = f"{diff.days} days, {diff.seconds//3600} hours, {(diff.seconds%3600)//60} minutes, and {diff.seconds%3600%60} seconds"
-                    else:
-                        strdiff = f'{diff.seconds//3600} hours, {(diff.seconds%3600)//60} minutes, and {diff.seconds%3600%60} seconds'
-                    d = {
-                        'start': status,
-                        'end': self.now().isoformat(),
-                        'time': strdiff,
-                    }
-                    deduction = self.calc_deduct(diff)
-
-                    #update return logs
-                    bikes_data = self.get_bikes()
-                    start_time = datetime.datetime.fromisoformat(bikes_data[bike_name]['status']).strftime('%Y/%m/%d, %H:%M:%S')
-                    end_time = self.now().strftime('%Y/%m/%d, %H:%M:%S')
-                    self.update_rental_log([bike_name,username,start_time,end_time,deduction])
-
-                    #update bike first, because bike uses user_data.bike_name
-                    bikes_data[bike_name]['status'] = 0
-                    bikes_data[bike_name]['username'] = ""
-                    self.update_bikes(bikes_data)
-
-                    #update user data
-                    log = user_data.get('log',[])
-                    log.append(d)
-                    user_data["status"] = None
-                    user_data["log"] = log
-                    user_data['bike_name'] = ''
-                    user_data['credits'] -= deduction
-                    user_data['finance'] = user_data.get('finance',[])
-                    f_log = {
-                        'type':'rental',
-                        'time':self.now().strftime("%Y/%m/%d, %H:%M:%S"),
-                        'credits':user_data.get('credits'),
-                        'spent': deduction,
-                        'remaining': user_data.get('credits') - deduction
-                    }
-                    user_data['finance'].append(f_log)
-                    super().update_user(user_data)
-
-                    admin_username = update.message.from_user.username
-                    context.bot.send_message(
-                        chat_id=user_data.get('chat_id'),
-                        text=f"An admin, @{admin_username} has returned your bike for you! Your total rental time is {strdiff}."
-                    )
-                    deduction_text = f"{deduction} credits was deducted. Remaining credits: {user_data['credits']}"
-                    deduction_text+= '\nIf you have any queries, please ask a comm member for help.'
-                    context.bot.send_message(
-                        chat_id=user_data.get('chat_id'),
-                        text=deduction_text
-                    )
-                    # Notify Admin group
-                    admin_text=f'[RENTAL - RETURN] \n@{update.message.from_user.username} returned {bike_name} at following time:\n{self.now().strftime("%Y/%m/%d, %H:%M:%S")}'
-                    admin_text+=f'\nThis return was force-returned by @{admin_username}.'
-                    admin_text+='\n'+deduction_text
-                    self.admin_log(update,context, admin_text)
-                else:
-                    context.bot.send_message(
-                        chat_id=update.effective_chat.id,
-                        text="Bike not on rent! Please try again"
-                    )
-
-            elif command == "setpin":
-                number = keywords.pop(0)
-                if bike['pin']!=number:
-                    bikes_data[bike_name]['oldpin'] = bike['pin']
-                    bikes_data[bike_name]['pin'] = number
-                    self.update_bikes(bikes_data)
-                    context.bot.send_message(
-                        chat_id=update.effective_chat.id,
-                        text=f"Pin for {bike_name} updated to {number}!"
-                    )
-                else:
-                    context.bot.send_message(
-                        chat_id=update.effective_chat.id,
-                        text=f'Old pin is the same as {number}!'
-                    )
-            elif command == "setstatus":
-                if not keywords:
-                    context.bot.send_message(
-                        chat_id=update.effective_chat.id,
-                        text="Please send a status update!"
-                    )
-                    return
-                status = ' '.join(keywords)
-                rented = bike.get('username',"")
-                if rented == "":
-                    if status.isnumeric(): #to reset the status to 0
-                        status=int(status)
-                    bikes_data[bike_name]['status'] = status
-                    self.update_bikes(bikes_data)
-                    context.bot.send_message(
-                        chat_id=update.effective_chat.id,
-                        text=f"Status for {bike_name} updated to {status}!"
-                    )
-                else:
-                    context.bot.send_message(
-                        chat_id=update.effective_chat.id,
-                        text=f'Bike is rented by {rented}!'
-                    )
-
-
-            elif "bike" in command:
-                bikes_data = self.get_bikes()
-                text= '\n'.join(f'{bike["name"]}  --  {bike["username"] or bike["status"] or EMOJI["tick"]} (Pin: {bike["pin"]})' for bike in bikes_data.values())
-                context.bot.send_message(
-                    chat_id=update.effective_chat.id,
-                    text=text)
-
-            elif command == "logs":
-                context.bot.send_document(
-                    chat_id=update.effective_chat.id,
-                    document=open('logs/rental.csv','rb'),
-                    filename="rental.csv",
-                    caption="Rental logs"
-                )
-                context.bot.send_document(
-                    chat_id=update.effective_chat.id,
-                    document=open('logs/report.csv','rb'),
-                    filename="report.csv",
-                    caption="Report logs"
-                )
-                context.bot.send_document(
-                    chat_id=update.effective_chat.id,
-                    document=open('logs/finance.csv','rb'),
-                    filename="finance.csv",
-                    caption="Finance logs"
-                )
-
-
-            else: # unrecognized command...
-                context.bot.send_message(
-                    chat_id=update.effective_chat.id,
-                    text='Unrecognized command. Try again. For more info, enter /admin')
-
-
-        except IndexError as e:
-            self.log_exception(e,"Error with handle_admin")
-            context.bot.send_message(
-                chat_id=update.effective_chat.id,
-                text=f"Sorry, too little info provided.\nPlease send more info after /{command}"
-                )
-        except ValueError as e:
-            self.log_exception(e,"Error with handle_admin")
-            context.bot.send_message(
-                chat_id=update.effective_chat.id,
-                text=f'Number entered, {number}, is not valid!')
-        except KeyError as e:
-            self.log_exception(e,"Error with handle_admin")
-            context.bot.send_message(
-                chat_id=update.effective_chat.id,
-                text=f'Could not find key {e} in dictionary! Please check database again')
-        except FileNotFoundError as e:
-            self.log_exception(e,"Error with handle_admin")
-            context.bot.send_message(
-                chat_id=update.effective_chat.id,
-                text='Hmm, file not found... Please raise a ticket with @fluffballz, along with what you sent.')
-        except Exception as e:
-            self.log_exception(e,"Error with handle_admin")
-            context.bot.send_message(
-                chat_id=update.effective_chat.id,
-                text=f'Failed, error is {e}\nPlease raise a ticket with @fluffballz, along with what you sent')
-
-
 if __name__=="__main__":
     newbot = OrcaBot(DEV_API_KEY, admin_group_id=DEV_ADMIN_GROUP_ID)
     newbot.main()
